Added ENS explorer compatibility check + bug fix
Quite a lot of refactor work was done on this commit, and I should have just commited each part separately. But I wanted to fix the bug I found because it was messing up my testing.

Bug Fix : basically stopped regex being applied to anything that wasn't a textNode, otherwise we would have attribute overwrites and other kinds of mess. In order for this to work I've created a tempory "slot" element that holds new elements, after regex is applied they are moved into the parent node.

Added docblocks